+++ src/app/components/attendees/AttendeesSettings.js
@@ -2,47 +2,65 @@ import React, { Component, Fragment } from "react";
 import bowser from "bowser";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import ReactTooltip from "react-tooltip";
 import VoxeetSDK from "@voxeet/voxeet-web-sdk";
 import Cookies from "./../../libs/Storage";
 import { Actions as InputManagerActions } from "../../actions/InputManagerActions";
 import { Actions as ConferenceActions } from "../../actions/ConferenceActions";
-import AttendeesParticipantVideo from "./AttendeesParticipantVideo";
+import { Actions as ControlsActions } from "../../actions/ControlsActions"
 import AttendeesSettingsVuMeterFromAudioLevel from "./AttendeesSettingsVuMeterFromAudioLevel";
-import AttendeesSettingsVuMeterFromMediaStream from "./AttendeesSettingsVuMeterFromMediaStream";
 import { strings } from "../../languages/localizedStrings";
-import { getVideoDeviceName } from "./../../libs/getVideoDeviceName";
-import {isIOS, isMobile, isElectron} from "./../../libs/browserDetection";
-import {Actions as ControlsActions} from "../../actions/ControlsActions";
-import {getUxKitContext} from "../../context";
+import { getVideoDeviceName } from "../../libs/getVideoDeviceName";
+import { isElectron, isIOS, isMobile } from "../../libs/browserDetection";
+import { getUxKitContext } from "../../context";
 
-var today = new Date();
+const today = new Date();
 today.setDate(today.getDate() + 365);
 const default_cookies_param = {
   path: "/",
   expires: today,
   secure: true,
-  sameSite: 'none'
+  sameSite: "none",
 };
 
-@connect(store => {
-  return {
-    inputManager: store.voxeet.inputManager,
-    controlsStore: store.voxeet.controls
-  };
-}, null, null, { context: getUxKitContext() })
+@connect(
+  (store) => {
+    return {
+      inputManager: store.voxeet.inputManager,
+      controlsStore: store.voxeet.controls,
+    };
+  },
+  null,
+  null,
+  { context: getUxKitContext() }
+)
 class AttendeesSettings extends Component {
   constructor(props) {
     super(props);
     // defaults
-    let maxVideoForwarding = ((this.props.controlsStore.maxVideoForwarding !== undefined) ? this.props.controlsStore.maxVideoForwarding : isMobile()?4:9);
-    let audioTransparentMode = ((this.props.controlsStore.audioTransparentMode !== undefined) ? this.props.controlsStore.audioTransparentMode : false);
-    let videoEnabled = ((this.props.controlsStore.videoEnabled !== undefined) ? this.props.controlsStore.videoEnabled : true);
-    let lowBandwidthMode = !videoEnabled && !maxVideoForwarding
-    let virtualBackgroundMode = ((this.props.controlsStore.virtualBackgroundMode !== undefined) ? this.props.controlsStore.virtualBackgroundMode : Cookies.get("virtualBackgroundMode"));
-    if(virtualBackgroundMode=='null')
-      virtualBackgroundMode = null;
-    let videoDenoise = ((this.props.controlsStore.videoDenoise !== undefined) ? this.props.controlsStore.videoDenoise : false);
+    let maxVideoForwarding =
+      this.props.controlsStore.maxVideoForwarding !== undefined
+        ? this.props.controlsStore.maxVideoForwarding
+        : isMobile()
+        ? 4
+        : 9;
+    let audioTransparentMode =
+      this.props.controlsStore.audioTransparentMode !== undefined
+        ? this.props.controlsStore.audioTransparentMode
+        : false;
+    let videoEnabled =
+      this.props.controlsStore.videoEnabled !== undefined
+        ? this.props.controlsStore.videoEnabled
+        : true;
+    let lowBandwidthMode = !videoEnabled && !maxVideoForwarding;
+    let virtualBackgroundMode =
+      this.props.controlsStore.virtualBackgroundMode !== undefined
+        ? this.props.controlsStore.virtualBackgroundMode
+        : Cookies.get("virtualBackgroundMode");
+    if (virtualBackgroundMode === "null") virtualBackgroundMode = null;
+    let videoDenoise =
+      this.props.controlsStore.videoDenoise !== undefined
+        ? this.props.controlsStore.videoDenoise
+        : false;
 
     this.state = {
       runningAnimation: false,
@@ -57,16 +75,20 @@ class AttendeesSettings extends Component {
       maxVideoForwarding: maxVideoForwarding,
       lowBandwidthMode: lowBandwidthMode,
       virtualBackgroundMode: virtualBackgroundMode,
-      videoDenoise: videoDenoise
+      videoDenoise: videoDenoise,
     };
     this.onAudioDeviceSelected = this.onAudioDeviceSelected.bind(this);
     this.setVideoDevice = this.setVideoDevice.bind(this);
     this.onOutputDeviceSelected = this.onOutputDeviceSelected.bind(this);
     this.onDeviceChange = this.onDeviceChange.bind(this);
-    this.handleChangeLowBandwidthMode = this.handleChangeLowBandwidthMode.bind(this);
-    this.onAudioTransparentModeChange = this.onAudioTransparentModeChange.bind(this);
-    this.handleMaxVideoForwardingChange = this.handleMaxVideoForwardingChange.bind(this);
-    this.onVirtualBackgroundModeChange = this.onVirtualBackgroundModeChange.bind(this);
+    this.handleChangeLowBandwidthMode =
+      this.handleChangeLowBandwidthMode.bind(this);
+    this.onAudioTransparentModeChange =
+      this.onAudioTransparentModeChange.bind(this);
+    this.handleMaxVideoForwardingChange =
+      this.handleMaxVideoForwardingChange.bind(this);
+    this.onVirtualBackgroundModeChange =
+      this.onVirtualBackgroundModeChange.bind(this);
     this.onVideoDenoiseChange = this.onVideoDenoiseChange.bind(this);
     this.maxVFTimer = null;
 
@@ -75,15 +97,20 @@ class AttendeesSettings extends Component {
 
   componentDidMount() {
     this.initDevices();
-    navigator.mediaDevices.addEventListener('devicechange', this.onDeviceChange);
+    navigator.mediaDevices.addEventListener(
+      "devicechange",
+      this.onDeviceChange
+    );
   }
 
-  static getDerivedStateFromProps(nextProps, prevState){
-    let stateUpdate = {attendeesSettingsOpened: nextProps.attendeesSettingsOpened};
+  static getDerivedStateFromProps(nextProps, prevState) {
+    let stateUpdate = {
+      attendeesSettingsOpened: nextProps.attendeesSettingsOpened,
+    };
     // Check if it should run animation
     if (
-        prevState.attendeesSettingsOpened == true &&
-        nextProps.attendeesSettingsOpened == false
+      prevState.attendeesSettingsOpened === true &&
+      nextProps.attendeesSettingsOpened === false
     ) {
       stateUpdate.runningAnimation = true;
     }
@@ -91,22 +118,18 @@ class AttendeesSettings extends Component {
   }
 
   componentDidUpdate(prevProps, prevState, snapshot) {
-    // Check if it should stop animation
     if (
-        prevProps.attendeesListOpened == true &&
-        this.props.attendeesListOpened == false
+      prevProps.attendeesListOpened === true &&
+      this.props.attendeesListOpened === false
     ) {
       setTimeout(() => {
         this.setState({ runningAnimation: false });
       }, 250);
     }
-  }
-
-  componentDidUpdate(prevProps, prevState) {
 
     if (
-        prevProps.attendeesSettingsOpened == true &&
-        this.props.attendeesSettingsOpened == false
+      prevProps.attendeesSettingsOpened === true &&
+      this.props.attendeesSettingsOpened === false
     ) {
       setTimeout(() => {
         this.setState({ runningAnimation: false });
@@ -115,16 +138,20 @@ class AttendeesSettings extends Component {
     //console.log(this.props.controlsStore.maxVideoForwarding, prevProps.controlsStore.maxVideoForwarding);
     if (
       this.props.controlsStore &&
-        prevProps.controlsStore.maxVideoForwarding !== this.props.controlsStore.maxVideoForwarding
+      prevProps.controlsStore.maxVideoForwarding !==
+        this.props.controlsStore.maxVideoForwarding
     ) {
       this.setState({
         maxVideoForwarding: this.props.controlsStore.maxVideoForwarding,
-        lowBandwidthMode: !this.props.controlsStore.videoEnabled && !this.props.controlsStore.maxVideoForwarding
+        lowBandwidthMode:
+          !this.props.controlsStore.videoEnabled &&
+          !this.props.controlsStore.maxVideoForwarding,
       });
     }
     if (
       this.props.controlsStore &&
-        prevProps.controlsStore.audioTransparentMode !== this.props.controlsStore.audioTransparentMode
+      prevProps.controlsStore.audioTransparentMode !==
+        this.props.controlsStore.audioTransparentMode
     ) {
       this.setState({
         audioTransparentMode: this.props.controlsStore.audioTransparentMode,
@@ -132,7 +159,8 @@ class AttendeesSettings extends Component {
     }
     if (
       this.props.controlsStore &&
-        prevProps.controlsStore.videoEnabled !== this.props.controlsStore.videoEnabled
+      prevProps.controlsStore.videoEnabled !==
+        this.props.controlsStore.videoEnabled
     ) {
       this.setState({
         videoEnabled: this.props.controlsStore.videoEnabled,
@@ -140,31 +168,48 @@ class AttendeesSettings extends Component {
     }
     if (
       this.props.controlsStore &&
-        prevProps.controlsStore.videoEnabled !== this.props.controlsStore.videoEnabled
+      prevProps.controlsStore.videoEnabled !==
+        this.props.controlsStore.videoEnabled
     ) {
-      this.setState({ lowBandwidthMode: !this.props.controlsStore.videoEnabled && !this.props.controlsStore.maxVideoForwarding });
+      this.setState({
+        lowBandwidthMode:
+          !this.props.controlsStore.videoEnabled &&
+          !this.props.controlsStore.maxVideoForwarding,
+      });
     }
 
     if (
       this.props.controlsStore &&
-        prevProps.controlsStore.virtualBackgroundMode !== this.props.controlsStore.virtualBackgroundMode
+      prevProps.controlsStore.virtualBackgroundMode !==
+        this.props.controlsStore.virtualBackgroundMode
     ) {
-      console.log('virtualBackgroundMode changed %s -> %s', prevProps.controlsStore.virtualBackgroundMode, this.props.controlsStore.virtualBackgroundMode)
-      this.setState({ virtualBackgroundMode: this.props.controlsStore.virtualBackgroundMode });
+      console.log(
+        "virtualBackgroundMode changed %s -> %s",
+        prevProps.controlsStore.virtualBackgroundMode,
+        this.props.controlsStore.virtualBackgroundMode
+      );
+      this.setState({
+        virtualBackgroundMode: this.props.controlsStore.virtualBackgroundMode,
+      });
     }
 
     if (
       this.props.controlsStore &&
-        prevProps.controlsStore.videoDenoise !== this.props.controlsStore.videoDenoise
+      prevProps.controlsStore.videoDenoise !==
+        this.props.controlsStore.videoDenoise
     ) {
-      console.log('videoDenoise changed %s -> %s', prevProps.controlsStore.videoDenoise, this.props.controlsStore.videoDenoise)
+      console.log(
+        "videoDenoise changed %s -> %s",
+        prevProps.controlsStore.videoDenoise,
+        this.props.controlsStore.videoDenoise
+      );
       this.setState({ videoDenoise: this.props.controlsStore.videoDenoise });
     }
 
     if (
-        this.props.controlsStore.videoEnabled !=
+      this.props.controlsStore.videoEnabled !==
         prevProps.controlsStore.videoEnabled ||
-        this.props.controlsStore.audioEnabled !=
+      this.props.controlsStore.audioEnabled !==
         prevProps.controlsStore.audioEnabled
     ) {
       this.initDevices();
@@ -172,164 +217,176 @@ class AttendeesSettings extends Component {
   }
 
   componentWillUnmount() {
-    navigator.mediaDevices.removeEventListener('devicechange', this.onDeviceChange);
+    navigator.mediaDevices.removeEventListener(
+      "devicechange",
+      this.onDeviceChange
+    );
   }
 
   onDeviceChange() {
     this.initDevices();
   }
 
-  onAudioTransparentModeChange() {
-    const { audioTransparentMode } = this.props.controlsStore;
-    this.props.dispatch(ConferenceActions.toggleAudioTransparentMode(!audioTransparentMode));
-  }
-
   initDevices() {
-    VoxeetSDK.mediaDevice.enumerateAudioDevices("input").then(devices => {
+    VoxeetSDK.mediaDevice
+      .enumerateAudioInputDevices()
+      .then((devices) => {
         this.setState({ audioDevices: devices });
-
-        // Pick out an audio input device:
-        // 1. use a current selected device
-        // 2. OR use a default device
-        // 3. OR use the first device from the device list
-        const deviceInfo =
-          devices.length && (
-            devices.find(e => e.deviceId === this.props.inputManager.currentAudioDevice)
-            || devices.find(e => e.deviceId === "default")
-            || devices[0]
-          );
-
-        // If the selected device is default, select it by using its proper deviceId
-        // (other than "default") but keep UI informed that the "default" is still selected.
-        if (deviceInfo.deviceId === "default") {
-          return this.setAudioDevice(
-            devices.find(e => e.groupId === deviceInfo.groupId
-                           && e.deviceId !== "default")
-              .deviceId,
-            "default");
+        const currentInputDevice = VoxeetSDK.mediaDevice.selectedAudioInputDevice;
+        if (currentInputDevice) {
+          return this.props.dispatch(InputManagerActions.inputAudioChange(currentInputDevice));
         }
 
-        return this.setAudioDevice(deviceInfo.deviceId);
       })
-      .catch(e => console.error("Initializing an audio input device failed.", e));
+      .catch((e) =>
+        console.error("Initializing an audio input device failed.", e)
+      );
 
-    VoxeetSDK.mediaDevice.enumerateAudioDevices("output").then(devices => {
+    VoxeetSDK.mediaDevice
+      .enumerateAudioOutputDevices()
+      .then((devices) => {
         this.setState({ outputDevices: devices });
-
-        // Pick out an audio output device:
-        // 1. use a current selected device
-        // 2. OR use a default device
-        // 3. OR use the first device from the device list
-        const deviceInfo =
-          devices.length && (
-            devices.find(e => e.deviceId === this.props.inputManager.currentOutputDevice)
-            || devices.find(e => e.deviceId === "default")
-            || devices[0]
-          );
-
-        return this.setOutputDevice(deviceInfo.deviceId);
-      })
-      .catch(e => console.error("Initializing an audio output device failed.", e));
-
-
-    VoxeetSDK.mediaDevice.enumerateVideoDevices().then(devices => {
-      if (this.props.inputManager.currentVideoDevice != "") {
-        let exist = false;
-        devices.map((device, i) => {
-          if (device.deviceId == this.props.inputManager.currentVideoDevice && device.deviceId != "") {
-            exist = true;
-          }
-        });
-        if (!exist && devices.length) {
-          let selected_device = devices.find(device => device.deviceId == 'default');
-          if (!selected_device)
-            selected_device = devices[0];
-          Cookies.set("camera", selected_device.deviceId, default_cookies_param);
-          getVideoDeviceName(selected_device.deviceId)
-            .then((isBackCamera) => {
-              this.props.dispatch(InputManagerActions.inputVideoChange(selected_device.deviceId, isBackCamera))
-            })
+        const currentOutputDevice = VoxeetSDK.mediaDevice.selectedAudioOutputDevice;
+        if (currentOutputDevice) {
+          return this.props.dispatch(InputManagerActions.outputAudioChange(currentOutputDevice));
         }
-      }
-      this.setState({
-        videoDevices: devices
-      });
-    })
-    .catch((e) => console.error(e));
+      })
+      .catch((e) =>
+        console.error("Initializing an audio output device failed.", e)
+      );
+
+    VoxeetSDK.mediaDevice
+      .enumerateVideoInputDevices()
+      .then((devices) => {
+        const currentVideoDevice = VoxeetSDK.mediaDevice.selectedVideoInputDevice;
+        if (currentVideoDevice) {
+          getVideoDeviceName(currentVideoDevice.deviceId).then(
+            (isBackCamera) => {
+              this.props.dispatch(
+                InputManagerActions.inputVideoChange(
+                  currentVideoDevice,
+                  isBackCamera
+                )
+              );
+            }
+          );
+        }
+        this.setState({
+          videoDevices: devices,
+        });
+      })
+      .catch((e) => console.error(e));
   }
 
   onOutputDeviceSelected(e) {
-    this.setOutputDevice(e.target.value).catch(e =>
+    const device =  e.target.value ? JSON.parse(e.target.value) : {};
+    this.setOutputDevice(device).catch((e) =>
       console.error("Selecting audio output device failed.", e)
     );
   }
 
-  setOutputDevice(deviceId) {
-    return VoxeetSDK.mediaDevice.selectAudioOutput(deviceId).then(() => {
-      Cookies.set("output", deviceId, default_cookies_param);
-      this.props.dispatch(InputManagerActions.outputAudioChange(deviceId));
+  setOutputDevice(device) {
+    return VoxeetSDK.mediaDevice.selectAudioOutput(device).then(() => {
+      Cookies.setDevice("output", device, default_cookies_param);
+      this.props.dispatch(InputManagerActions.outputAudioChange(device));
     });
   }
 
   onAudioDeviceSelected(e) {
-    this.setAudioDevice(e.target.value).catch(e =>
+    const device =  e.target.value ? JSON.parse(e.target.value) : {};
+    this.setAudioDevice(device).catch((e) =>
       console.error("Selecting audio input device failed.", e)
     );
   }
 
-  setAudioDevice(deviceId, guiDeviceId = deviceId) {
-    return VoxeetSDK.mediaDevice.selectAudioInput(deviceId).then(() => {
+  setAudioDevice(device) {
+    return VoxeetSDK.mediaDevice.selectAudioInput(device).then(() => {
       if (this.props.microphoneMuted) {
         VoxeetSDK.conference
           .mute(VoxeetSDK.session.participant, true)
-          .catch((e) => console.warn("Muting a new selected input device failed.", e));
+          .catch((e) =>
+            console.warn("Muting a new selected input device failed.", e)
+          );
       }
 
-      Cookies.set("input", guiDeviceId, default_cookies_param);
-      this.props.dispatch(InputManagerActions.inputAudioChange(guiDeviceId));
+      Cookies.setDevice("input", device, default_cookies_param);
+      this.props.dispatch(InputManagerActions.inputAudioChange(device));
     });
   }
 
   setVideoDevice(e) {
     const { videoEnabled } = this.props;
-    const deviceId = e.target.value;
+    const device =  e.target.value ? JSON.parse(e.target.value) : {};
     if (videoEnabled) {
-      VoxeetSDK.mediaDevice.selectVideoInput(deviceId);
+      VoxeetSDK.mediaDevice.selectVideoInput(device);
     }
-    Cookies.set("camera", deviceId, default_cookies_param);
-    getVideoDeviceName(deviceId)
-      .then((isBackCamera, currentVideoDevice) => {
-        this.props.dispatch(InputManagerActions.inputVideoChange(deviceId, isBackCamera))
-      })
+    Cookies.setDevice("camera", device, default_cookies_param);
+    getVideoDeviceName(device.deviceId).then(
+      (isBackCamera, currentVideoDevice) => {
+        this.props.dispatch(
+          InputManagerActions.inputVideoChange(device, isBackCamera)
+        );
+      }
+    );
   }
 
   onAudioTransparentModeChange() {
     const { audioTransparentMode } = this.props.controlsStore;
-    this.setState({
-      audioTransparentMode: !audioTransparentMode
-    }, () => {
-      Cookies.set('audioTransparentMode', this.state.audioTransparentMode, default_cookies_param);
-      this.props.dispatch(ConferenceActions.setAudioTransparentMode(this.state.audioTransparentMode));
-    })
+    this.setState(
+      {
+        audioTransparentMode: !audioTransparentMode,
+      },
+      () => {
+        Cookies.set(
+          "audioTransparentMode",
+          this.state.audioTransparentMode,
+          default_cookies_param
+        );
+        this.props.dispatch(
+          ControlsActions.toggleAudioTransparentMode()
+        );
+        this.props.dispatch(
+          ConferenceActions.setAudioTransparentMode(
+            this.state.audioTransparentMode
+          )
+        );
+      }
+    );
   }
 
   onVirtualBackgroundModeChange(mode) {
-    console.log('onVirtualBackgroundModeChange', mode);
-    this.setState({
-      virtualBackgroundMode: mode!==this.state.virtualBackgroundMode?mode:null
-    }, () => {
-      console.log('about to call ConferenceActions.setVirtualBackgroundMode', this.state.virtualBackgroundMode);
-      this.props.dispatch(ConferenceActions.setVirtualBackgroundMode(this.state.virtualBackgroundMode));
-    });
+    console.log("onVirtualBackgroundModeChange", mode);
+    this.setState(
+      {
+        virtualBackgroundMode:
+          mode !== this.state.virtualBackgroundMode ? mode : null,
+      },
+      () => {
+        console.log(
+          "about to call ConferenceActions.setVirtualBackgroundMode",
+          this.state.virtualBackgroundMode
+        );
+        this.props.dispatch(
+          ConferenceActions.setVirtualBackgroundMode(
+            this.state.virtualBackgroundMode
+          )
+        );
+      }
+    );
   }
 
   onVideoDenoiseChange() {
     const { videoDenoise } = this.props.controlsStore;
-    this.setState({
-      videoDenoise: !videoDenoise
-    }, () => {
-      this.props.dispatch(ConferenceActions.setVideoDenoise(this.state.videoDenoise));
-    });
+    this.setState(
+      {
+        videoDenoise: !videoDenoise,
+      },
+      () => {
+        this.props.dispatch(
+          ConferenceActions.setVideoDenoise(this.state.videoDenoise)
+        );
+      }
+    );
   }
 
   handleChangeLowBandwidthMode(event) {
@@ -338,21 +395,35 @@ class AttendeesSettings extends Component {
       this.maxVFTimer = null;
     }
     const low_bandwidth = event.target.checked;
-    let maxVideoForwarding = low_bandwidth?0:(Cookies.get("maxVideoForwarding")!==undefined?Cookies.get("maxVideoForwarding"):(isMobile()?4:9));
-    if(typeof maxVideoForwarding === 'string' || maxVideoForwarding instanceof String)
+    let maxVideoForwarding = low_bandwidth
+      ? 0
+      : Cookies.get("maxVideoForwarding") !== undefined
+      ? Cookies.get("maxVideoForwarding")
+      : isMobile()
+      ? 4
+      : 9;
+    if (
+      typeof maxVideoForwarding === "string" ||
+      maxVideoForwarding instanceof String
+    )
       maxVideoForwarding = parseInt(maxVideoForwarding);
 
-    if(low_bandwidth) {
+    if (low_bandwidth) {
       // Get current videoEnabled status
-      let videoEnabled = ((this.props.controlsStore.videoEnabled !== undefined) ? this.props.controlsStore.videoEnabled : true);
+      let videoEnabled =
+        this.props.controlsStore.videoEnabled !== undefined
+          ? this.props.controlsStore.videoEnabled
+          : true;
       // Cookies.set('videoEnabled', false, default_cookies_param);
-      if(videoEnabled) {
+      if (videoEnabled) {
         // Disable if enabled, don't change if already disabled
         this.props.dispatch(ConferenceActions.toggleVideo(true));
       }
       // Don't change maxVideoForwarding value in cookies
       // Cookies.set("maxVideoForwarding", 0, default_cookies_param);
-      this.props.dispatch(ConferenceActions.setMaxVideoForwarding(maxVideoForwarding));
+      this.props.dispatch(
+        ConferenceActions.setMaxVideoForwarding(maxVideoForwarding)
+      );
     } else {
       // Don't enable video, don't touch a value in cookies
       // Cookies.set('videoEnabled', true, default_cookies_param);
@@ -360,7 +431,9 @@ class AttendeesSettings extends Component {
       // Don't change maxVideoForwarding value in cookies
       // Cookies.set("maxVideoForwarding", maxVideoForwarding, default_cookies_param);
       // Revert to value stored in cookies
-      this.props.dispatch(ConferenceActions.setMaxVideoForwarding(maxVideoForwarding));
+      this.props.dispatch(
+        ConferenceActions.setMaxVideoForwarding(maxVideoForwarding)
+      );
     }
   }
 
@@ -372,7 +445,9 @@ class AttendeesSettings extends Component {
     let num;
     const maxVF = event.target.value;
 
-    try { num = parseInt(maxVF) } catch (e) { };
+    try {
+      num = parseInt(maxVF);
+    } catch (e) {}
     if (num !== undefined && !isNaN(num)) {
       this.setState({ maxVideoForwarding: num }, () => {
         this.maxVFTimer = setTimeout(() => {
@@ -383,31 +458,35 @@ class AttendeesSettings extends Component {
           this.maxVFTimer = null;
         }, 1500);
       });
-    }
-    else {
-      this.setState({ maxVideoForwarding: '' });
+    } else {
+      this.setState({ maxVideoForwarding: "" });
     }
   }
 
   render() {
-    const { lowBandwidthMode, maxVideoForwarding, audioTransparentMode, virtualBackgroundMode, videoEnabled, videoDenoise } = this.state;
+    const {
+      lowBandwidthMode,
+      maxVideoForwarding,
+      audioTransparentMode,
+      virtualBackgroundMode,
+      videoEnabled,
+      videoDenoise,
+    } = this.state;
     //const { audioTransparentMode } = this.props.controlsStore;
 
-    const { attendeesSettingsOpened, isListener, dolbyVoiceEnabled } = this.props;
-    const MAX_MAXVF = isMobile()?4:16;
-    const {
-      currentAudioDevice,
-      currentVideoDevice,
-      currentOutputDevice
-    } = this.props.inputManager;
+    const { attendeesSettingsOpened, isListener, dolbyVoiceEnabled } =
+      this.props;
+    const MAX_MAXVF = isMobile() ? 4 : 16;
+    const { currentAudioDevice, currentVideoDevice, currentOutputDevice } =
+      this.props.inputManager;
     return (
       <div
         className={
           this.state.runningAnimation
             ? "attendees-settings attendees-settings-out"
             : attendeesSettingsOpened
-              ? "attendees-settings"
-              : "attendees-settings-hidden"
+            ? "attendees-settings"
+            : "attendees-settings-hidden"
         }
       >
         <div className="attendees-settings-header">
@@ -422,32 +501,32 @@ class AttendeesSettings extends Component {
                   {/* <label htmlFor="output">Sound Output</label> */}
                   <select
                     name="output"
-                    value={currentOutputDevice}
+                    value={JSON.stringify(currentOutputDevice)}
                     className="form-control select-audio-output"
                     onChange={this.onOutputDeviceSelected}
                     disabled={false}
                   >
                     {this.state.outputDevices.map((device, i) => (
-                      <option key={i} value={device.deviceId}>
+                      <option key={i} value={JSON.stringify(device)}>
                         {device.label}
                       </option>
                     ))}
                   </select>
                 </div>
               )}
-              {!isListener &&
-                (<Fragment>
+              {!isListener && (
+                <Fragment>
                   <div className="form-group last">
                     {/* <label htmlFor="video">Camera</label> */}
                     <select
                       name="video"
-                      value={currentVideoDevice}
+                      value={JSON.stringify(currentVideoDevice)}
                       className="form-control select-video-device"
                       onChange={this.setVideoDevice}
                       disabled={false}
                     >
                       {this.state.videoDevices.map((device, i) => (
-                        <option key={i} value={device.deviceId}>
+                        <option key={i} value={JSON.stringify(device)}>
                           {device.label}
                         </option>
                       ))}
@@ -457,101 +536,126 @@ class AttendeesSettings extends Component {
                     {/* <label htmlFor="video">Microphone</label> */}
                     <select
                       name="audio"
-                      value={currentAudioDevice}
+                      value={JSON.stringify(currentAudioDevice)}
                       className="form-control select-audio-input"
                       onChange={this.onAudioDeviceSelected}
                     >
                       {this.state.audioDevices.map((device, i) => (
-                        <option key={i} value={device.deviceId}>
+                        <option key={i} value={JSON.stringify(device)}>
                           {device.label}
                         </option>
                       ))}
                     </select>
                   </div>
                   <div className="form-group">
-                    {(this.isIOS || bowser.safari) ?
-                      <AttendeesSettingsVuMeterFromAudioLevel maxLevel={21} /> :
-                      <AttendeesSettingsVuMeterFromMediaStream maxLevel={21} />}
+                    {<AttendeesSettingsVuMeterFromAudioLevel maxLevel={21} />}
                   </div>
-                  {dolbyVoiceEnabled && <div className="form-group switch-enable">
-                    <div className='switch-mode'>
-                      <input
-                        id="audioTransparentMode"
-                        name="audioTransparentMode"
-                        type="checkbox"
-                        onChange={this.onAudioTransparentModeChange}
-                        checked={audioTransparentMode}
-                      />
-                      <label htmlFor="audioTransparentMode">
-                        {strings.audioTransparentMode}
-                      </label>
+                  {dolbyVoiceEnabled && (
+                    <div className="form-group switch-enable">
+                      <div className="switch-mode">
+                        <input
+                          id="audioTransparentMode"
+                          name="audioTransparentMode"
+                          type="checkbox"
+                          onChange={this.onAudioTransparentModeChange}
+                          checked={audioTransparentMode}
+                        />
+                        <label htmlFor="audioTransparentMode">
+                          {strings.audioTransparentMode}
+                        </label>
+                      </div>
                     </div>
-                  </div>}
-                  </Fragment>
-                )
-              }
-                <div className="form-group switch-enable">
-                  <div className='switch-mode'>
-                    <input
-                      id="lowBandwidthMode"
-                      name="lowBandwidthMode"
-                      type="checkbox"
-                      onChange={this.handleChangeLowBandwidthMode}
-                      checked={lowBandwidthMode}
-                    />
-                    <label htmlFor="lowBandwidthMode">
-                      {strings.lowBandwidthMode}
-                    </label>
-                  </div>
-                </div>
-                {isElectron() && <div className={`form-group switch-enable ${!videoEnabled ? 'disabled-form' : ''}`}>
-                  <div className='switch-mode'>
-                    <input
-                        id="vbModeBokeh"
-                        name="vbModeBokeh"
-                        type="checkbox"
-                        onChange={() => this.onVirtualBackgroundModeChange('bokeh')}
-                        checked={virtualBackgroundMode == 'bokeh'}
-                    />
-                    <label htmlFor="vbModeBokeh">
-                      {strings.bokehMode}
-                    </label>
-                  </div>
-                </div>}
-                {isElectron() && <div className={`form-group switch-enable ${!videoEnabled ? 'disabled-form' : ''}`}>
-                  <div className='switch-mode'>
-                    <input
-                        id="videoDenoise"
-                        name="videoDenoise"
-                        type="checkbox"
-                        onChange={this.onVideoDenoiseChange}
-                        checked={videoDenoise}
-                    />
-                    <label htmlFor="videoDenoise">
-                      {strings.videoDenoise}
-                    </label>
-                  </div>
-                </div>}
-                <div className={`form-group switch-enable maxVideoForwarding ${lowBandwidthMode ? 'disabled-form' : ''}`}>
-                  <div className='input-wrapper'>
-                    <div className='input-value'>0</div>
-                    <input
-                      type="range"
-                      style={{ background: `linear-gradient(to right, #00afef 0%, #00afef ${(100 / MAX_MAXVF) * (!lowBandwidthMode ? maxVideoForwarding : 0)}%, #fff ${(100 / MAX_MAXVF) * (!lowBandwidthMode ? maxVideoForwarding : 0)}%, #fff 100%)` }}
-                      id="maxVideoForwarding"
-                      name="maxVideoForwarding"
-                      min={0}
-                      max={MAX_MAXVF}
-                      step={1}
-                      onChange={this.handleMaxVideoForwardingChange}
-                      value={!lowBandwidthMode ? maxVideoForwarding : 0}
-                    />
-                    <div className='input-value'>{MAX_MAXVF}</div>
-                  </div>
-                  <label htmlFor="maxVideoForwarding">
-                    <div className='maxVideoForwardingValue'>{`${strings.showVideoParticipants1} ${!lowBandwidthMode ? maxVideoForwarding : 0} ${strings.showVideoParticipants2}`}</div>
+                  )}
+                </Fragment>
+              )}
+              <div className="form-group switch-enable">
+                <div className="switch-mode">
+                  <input
+                    id="lowBandwidthMode"
+                    name="lowBandwidthMode"
+                    type="checkbox"
+                    onChange={this.handleChangeLowBandwidthMode}
+                    checked={lowBandwidthMode}
+                  />
+                  <label htmlFor="lowBandwidthMode">
+                    {strings.lowBandwidthMode}
                   </label>
                 </div>
+              </div>
+              {isElectron() && (
+                <div
+                  className={`form-group switch-enable ${
+                    !videoEnabled ? "disabled-form" : ""
+                  }`}
+                >
+                  <div className="switch-mode">
+                    <input
+                      id="vbModeBokeh"
+                      name="vbModeBokeh"
+                      type="checkbox"
+                      onChange={() =>
+                        this.onVirtualBackgroundModeChange("bokeh")
+                      }
+                      checked={virtualBackgroundMode === "bokeh"}
+                    />
+                    <label htmlFor="vbModeBokeh">{strings.bokehMode}</label>
+                  </div>
+                </div>
+              )}
+              {isElectron() && (
+                <div
+                  className={`form-group switch-enable ${
+                    !videoEnabled ? "disabled-form" : ""
+                  }`}
+                >
+                  <div className="switch-mode">
+                    <input
+                      id="videoDenoise"
+                      name="videoDenoise"
+                      type="checkbox"
+                      onChange={this.onVideoDenoiseChange}
+                      checked={videoDenoise}
+                    />
+                    <label htmlFor="videoDenoise">{strings.videoDenoise}</label>
+                  </div>
+                </div>
+              )}
+              <div
+                className={`form-group switch-enable maxVideoForwarding ${
+                  lowBandwidthMode ? "disabled-form" : ""
+                }`}
+              >
+                <div className="input-wrapper">
+                  <div className="input-value">0</div>
+                  <input
+                    type="range"
+                    style={{
+                      background: `linear-gradient(to right, #00afef 0%, #00afef ${
+                        (100 / MAX_MAXVF) *
+                        (!lowBandwidthMode ? maxVideoForwarding : 0)
+                      }%, #fff ${
+                        (100 / MAX_MAXVF) *
+                        (!lowBandwidthMode ? maxVideoForwarding : 0)
+                      }%, #fff 100%)`,
+                    }}
+                    id="maxVideoForwarding"
+                    name="maxVideoForwarding"
+                    min={0}
+                    max={MAX_MAXVF}
+                    step={1}
+                    onChange={this.handleMaxVideoForwardingChange}
+                    value={!lowBandwidthMode ? maxVideoForwarding : 0}
+                  />
+                  <div className="input-value">{MAX_MAXVF}</div>
+                </div>
+                <label htmlFor="maxVideoForwarding">
+                  <div className="maxVideoForwardingValue">{`${
+                    strings.showVideoParticipants1
+                  } ${!lowBandwidthMode ? maxVideoForwarding : 0} ${
+                    strings.showVideoParticipants2
+                  }`}</div>
+                </label>
+              </div>
               <div className="hint-text">
                 <p>{strings.problemSettings}</p>
                 <p>{strings.saveSettings}</p>