+++ src/app/components/attendees/modes/index.js
@@ -2,5 +2,6 @@ export { default as List } from "./List";
 export { default as ListWidget } from "./ListWidget";
 export { default as Speakers } from "./Speakers";
 export { default as Tiles } from "./Tiles";
+export { default as MeasuredTiles } from "./MeasuredTiles";
 export { default as View3D } from "./View3D";
 export { default as ToggleModeButton } from "./ToggleModeButton";