+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 ## Changelog
 
+### 3.5.0
+
+#### Features
+
+- Support for [SDK 3.5.0](https://github.com/voxeet/voxeet-sdk-web/releases/tag/3.5.0) that supports the Dolby Voice Codec in Dolby Voice conferences to improve audio processing and offer higher audio quality.
+
+#### Bugs
+
+- Fixed an issue where a file presentation started by a participant without audio and video was not visible to the rest of conference participants ([#214](https://github.com/voxeet/voxeet-uxkit-react/issues/214))
+- Fixed an issue where UXKit did not use the customLocalizedStrings property ([#229](https://github.com/voxeet/voxeet-uxkit-react/issues/229))
+
+### 3.3.0
+
+#### Features
+
+- Support for SDK 3.4 and [Spatial Audio](https://docs.dolby.io/communications-apis/docs/guides-spatial-audio) with Audio/Video congruence
+- Support for React 17
+
+### 3.2.4
+
+#### Bug fixes
+
+- Fixed an issue where media track remains open after the end of the call
+
+### 3.2.3
+
+#### Bug fixes
+
+- Fixed an issue where listeners and participants without streams are not visible in the conference
+- Fixed an issue where the switch for bokeh mode is visible in web browsers
+
 ### 3.2.2
 
 #### Bug fixes